feat: :wrench: Ajuste de confs.
